consensus: update processPeerUpdate for new peers
